netphantom/requirements.txt
rich>=13.0
dnspython>=2.4
python-whois>=0.9
